console/requirements.txt
# app
requests
django-environ

# utility
setuptools-scm

# opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-urllib3

# exporter
opentelemetry-exporter-otlp-proto-grpc
